requirements.txt
bittensor
pytest-cov==5.0.0
fourier @ git+https://github.com/apollozkp/fourier.git@move_fft
